Move RefreshPrompt to new location and move SW refresh code
