+++ www/src/js/views/settings/RefreshPrompt.jsx
@@ -2,10 +2,11 @@
 import React, { PureComponent } from 'react';
 import classnames from 'classnames';
 import { connect } from 'react-redux';
+
 import type { State } from 'reducers';
-import { getRegistration } from 'bootstrapping/service-worker';
+import { updateServiceWorker } from 'bootstrapping/service-worker';
 import { Refresh } from 'views/components/icons';
-import styles from './Announcements.scss';
+import styles from 'views/components/notfications/Announcements.scss';
 
 type Props = {
   showPrompt: boolean,
@@ -26,20 +27,7 @@ class RefreshPrompt extends PureComponent<Props> {
           <p>Please refresh the page to get the latest version.</p>
         </div>
 
-        <button
-          className="btn btn-success"
-          type="button"
-          onClick={() => {
-            const registration = getRegistration();
-            if (!registration || !registration.waiting) {
-              // Just to ensure registration.waiting is available before
-              // calling postMessage()
-              return;
-            }
-
-            registration.waiting.postMessage('skipWaiting');
-          }}
-        >
+        <button className="btn btn-success" type="button" onClick={updateServiceWorker}>
           Refresh page
         </button>
       </div>
