+++ www/src/js/views/timetable/TimetableContent.jsx
@@ -37,7 +37,7 @@ import ModulesSelectContainer from 'views/timetable/ModulesSelectContainer';
 import CorsNotification from 'views/components/cors-info/CorsNotification';
 import Announcements from 'views/components/notfications/Announcements';
 import Title from 'views/components/Title';
-import RefreshPrompt from 'views/components/notfications/RefreshPrompt';
+import RefreshPrompt from 'views/settings/RefreshPrompt';
 import NoLessonWarning from 'views/timetable/NoLessonWarning';
 import Timetable from './Timetable';
 import TimetableActions from './TimetableActions';
